+++ tango/integrations/torch/train_callback.py
@@ -2,11 +2,13 @@ from pathlib import Path
 from typing import Any, Dict, List, Optional
 
 import torch
+from overrides import overrides
 
-from tango.common.exceptions import TangoError
+from tango.common.dataset_dict import DatasetDictBase
 from tango.common.registrable import Registrable
 
 from .data import DataLoader
+from .exceptions import StopEarly
 from .model import Model
 from .optim import LRScheduler, Optimizer
 
@@ -32,24 +34,34 @@ class TrainCallback(Registrable):
         work_dir: Path,
         model: Model,
         optimizer: Optimizer,
+        dataset_dict: DatasetDictBase,
         train_dataloader: DataLoader,
+        train_steps: int,
+        validation_steps: Optional[int] = None,
         validation_dataloader: Optional[DataLoader] = None,
         lr_scheduler: Optional[LRScheduler] = None,
         is_local_main_process: bool = True,
         world_size: int = 1,
         worker_id: int = 0,
         device: torch.device = torch.device("cpu"),
+        val_metric_name: str = "loss",
+        minimize_val_metric: bool = True,
     ) -> None:
         self.work_dir = work_dir
         self.model = model
         self.optimizer = optimizer
+        self.dataset_dict = dataset_dict
         self.train_dataloader = train_dataloader
+        self.train_steps = train_steps
+        self.validation_steps = validation_steps
         self.validation_dataloader = validation_dataloader
         self.lr_scheduler = lr_scheduler
         self.is_local_main_process = is_local_main_process
         self.worker_id = worker_id
         self.world_size = world_size
         self.device = device
+        self.val_metric_name = val_metric_name
+        self.minimize_val_metric = minimize_val_metric
 
     def state_dict(self) -> Dict[str, Any]:
         """
@@ -139,18 +151,32 @@ class TrainCallback(Registrable):
         """
         pass
 
-    def post_val_loop(self, step: int, val_metric_name: str, val_metric: float) -> None:
+    def post_val_loop(self, step: int, val_metric: float, best_val_metric: float) -> None:
         """
         Called right after the validation loop finishes.
         """
         pass
 
 
-class StopEarly(TangoError):
+@TrainCallback.register("torch::stop_early")
+class StopEarlyCallback(TrainCallback):
     """
-    Callbacks can raise this exception to stop training early without crashing.
+    A :class:`TrainCallback` for early stopping. Training is stopped early after
+    ``patience`` steps without an improvement to the validation metric.
 
-    .. important::
-        During distributed training all workers must raise this exception at the same point
-        in the training loop, otherwise there will be a deadlock.
+    .. tip::
+
+        Registered as a :class:`TrainCallback` under the name "torch::stop_early".
     """
+
+    def __init__(self, *args, patience: int = 10000, **kwargs) -> None:
+        super().__init__(*args, **kwargs)
+        self.patience = patience
+        self.best_step = 0
+
+    @overrides
+    def post_val_loop(self, step: int, val_metric: float, best_val_metric: float) -> None:
+        if val_metric == best_val_metric:
+            self.best_step = step
+        elif step > self.best_step + self.patience:
+            raise StopEarly