+++ CHANGELOG.md
@@ -7,6 +7,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v0.3.5](https://github.com/allenai/tango/releases/tag/v0.3.5) - 2021-11-05
+
+### Fixed
+
+- Fixed a bug in `FromParams` where the "type" parameter was ignored in some cases
+  where the `Registrable` base class did not directly inherit from `Registrable`.
+
+## [v0.3.4](https://github.com/allenai/tango/releases/tag/v0.3.4) - 2021-11-04
+
+### Added
+
+- Added `StopEarlyCallback`, a `TorchTrainCallback` for early stopping.
+- Added parameter `remove_stale_checkpoints` to `TorchTrainStep`.
+
+### Changed
+
+- Minor changes to `TorchTrainCallback` interface.
+- Weights & Biases `TorchTrainCallback` now logs best validation metric score.
+
+## [v0.3.3](https://github.com/allenai/tango/releases/tag/v0.3.3) - 2021-11-04
+
 ### Added
 
 - Added support for PEP 604 in `FromParams`, i.e. writing union types as "X | Y" instead of "Union[X, Y]".
@@ -17,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed a bug in `FromParams` where non-`FromParams` class parameters were not instantiated
   properly (or at all).
+- Fixed a bug in `FromParams` where kwargs were not passed on from a wrapper class to the wrapped class.
 - Fixed small bug where some errors from git would be printed when executor metadata is created
   outside of a git repository.